+++ ts/features/barcode/screens/BarcodeScanScreen.tsx
@@ -146,7 +146,7 @@ const BarcodeScanScreen = () => {
         };
         NavigationService.dispatchNavigationAction(
           CommonActions.navigate(ITW_ROUTES.MAIN, {
-            screen: ITW_ROUTES.PRESENTATION.PID.REMOTE.INIT,
+            screen: ITW_ROUTES.PRESENTATION.PID.REMOTE.CHECKS,
             params
           })
         );