+++ ts/features/it-wallet/screens/issuing/pid/cie/ItwCiePinScreen.tsx
@@ -17,29 +17,29 @@ import {
 } from "react-native";
 import { connect } from "react-redux";
 import { IOColors, VSpacer } from "@pagopa/io-app-design-system";
-import ButtonDefaultOpacity from "../../../../../components/ButtonDefaultOpacity";
-import CiePinpad from "../../../../../components/CiePinpad";
-import { Link } from "../../../../../components/core/typography/Link";
-import { ScreenContentHeader } from "../../../../../components/screens/ScreenContentHeader";
-import TopScreenComponent from "../../../../../components/screens/TopScreenComponent";
-import FooterWithButtons from "../../../../../components/ui/FooterWithButtons";
+import ButtonDefaultOpacity from "../../../../../../components/ButtonDefaultOpacity";
+import CiePinpad from "../../../../../../components/CiePinpad";
+import { Link } from "../../../../../../components/core/typography/Link";
+import { ScreenContentHeader } from "../../../../../../components/screens/ScreenContentHeader";
+import TopScreenComponent from "../../../../../../components/screens/TopScreenComponent";
+import FooterWithButtons from "../../../../../../components/ui/FooterWithButtons";
 import {
   BottomTopAnimation,
   LightModalContext
-} from "../../../../../components/ui/LightModal";
-import Markdown from "../../../../../components/ui/Markdown";
-import I18n from "../../../../../i18n";
-import { IOStackNavigationProp } from "../../../../../navigation/params/AppParamsList";
-import { Dispatch, ReduxProps } from "../../../../../store/actions/types";
-import variables from "../../../../../theme/variables";
-import { setAccessibilityFocus } from "../../../../../utils/accessibility";
-import { useLegacyIOBottomSheetModal } from "../../../../../utils/hooks/bottomSheet";
-import { openWebUrl } from "../../../../../utils/url";
-import { pinPukHelpUrl } from "../../../../../config";
-import { ItwParamsList } from "../../../navigation/ItwParamsList";
-import { ITW_ROUTES } from "../../../navigation/ItwRoutes";
-import { CieRequestAuthenticationOverlay } from "../../../components/cie/CieRequestAuthenticationOverlay";
-import { nfcIsEnabled } from "../../../store/actions/itwCieActions";
+} from "../../../../../../components/ui/LightModal";
+import Markdown from "../../../../../../components/ui/Markdown";
+import I18n from "../../../../../../i18n";
+import { IOStackNavigationProp } from "../../../../../../navigation/params/AppParamsList";
+import { Dispatch, ReduxProps } from "../../../../../../store/actions/types";
+import variables from "../../../../../../theme/variables";
+import { setAccessibilityFocus } from "../../../../../../utils/accessibility";
+import { useLegacyIOBottomSheetModal } from "../../../../../../utils/hooks/bottomSheet";
+import { openWebUrl } from "../../../../../../utils/url";
+import { pinPukHelpUrl } from "../../../../../../config";
+import { ItwParamsList } from "../../../../navigation/ItwParamsList";
+import { ITW_ROUTES } from "../../../../navigation/ItwRoutes";
+import { CieRequestAuthenticationOverlay } from "../../../../components/cie/CieRequestAuthenticationOverlay";
+import { nfcIsEnabled } from "../../../../store/actions/itwCieActions";
 
 const mapDispatchToProps = (dispatch: Dispatch) => ({
   requestNfcEnabledCheck: () => dispatch(nfcIsEnabled.request())
@@ -153,7 +153,7 @@ const ItwCiePinScreen: React.FC<Props> = props => {
       <ScrollView>
         <ScreenContentHeader
           title={I18n.t("authentication.cie.pin.pinCardTitle")}
-          rasterIcon={require("../../../../../../img/icons/icon_insert_cie_pin.png")}
+          rasterIcon={require("../../../../../../../img/icons/icon_insert_cie_pin.png")}
           subtitle={I18n.t("authentication.cie.pin.subtitleHelp")}
           subtitleLink={
             <Link onPress={present}>