+++ ts/features/it-wallet/screens/presentation/remote/credential/ItwPrCredentialDataScreen.tsx
@@ -45,7 +45,7 @@ type ContentViewParams = {
  * This screen displays the information about the credential that is going to be shared
  * with the issuer.
  */
-const ItwPresentationDataScreen = () => {
+const ItwPrCredentialDataScreen = () => {
   const decodedPid = useIOSelector(itwDecodedPidValueSelector);
   const navigation = useNavigation<IOStackNavigationProp<ItwParamsList>>();
   const rpMock = getRpMock();
@@ -247,4 +247,4 @@ const styles = StyleSheet.create({
   }
 });
 
-export default ItwPresentationDataScreen;
+export default ItwPrCredentialDataScreen;